after app.py simplification to move to only async workflow support
