+++ nifi_chat_ui/app.py
@@ -126,7 +126,6 @@ def get_mcp_client() -> MCPClient:
 logger.info("New modular LLM architecture initialized (lazy loading)")
 
 # --- Constants ---
-MAX_LOOP_ITERATIONS = 10 # Safety break for the execution loop
 SYSTEM_PROMPT_FILE = "nifi_chat_ui/system_prompt.md"
 
 # --- Session State Initialization ---
@@ -295,71 +294,60 @@ with st.sidebar:
     # --- Workflow Configuration --- 
     st.markdown("---") # Add separator
     
-    # Initialize workflow session state
-    if "workflow_execution_mode" not in st.session_state:
-        st.session_state.workflow_execution_mode = "unguided"
+        # Initialize workflow session state
     if "selected_workflow" not in st.session_state:
         st.session_state.selected_workflow = None
     if "available_workflows" not in st.session_state:
         st.session_state.available_workflows = []
     
-
-    
-    # Default to Guided mode (no selection needed)
-    st.session_state.workflow_execution_mode = "guided"
-    
-    # === WORKFLOW SELECTION (for guided mode) ===
-    if st.session_state.workflow_execution_mode == "guided":
-        try:
-            # Get available workflows from the workflow registry
-            from nifi_mcp_server.workflows.registry import get_workflow_registry
-            registry = get_workflow_registry()
-            workflows = registry.list_workflows(enabled_only=True)
+    # === WORKFLOW SELECTION ===
+    try:
+        # Get available workflows from the workflow registry
+        from nifi_mcp_server.workflows.registry import get_workflow_registry
+        registry = get_workflow_registry()
+        workflows = registry.list_workflows(enabled_only=True)
+        
+        if workflows:
+            # Create workflow options for the selectbox
+            workflow_options = {}
+            for workflow in workflows:
+                display_name = f"{workflow.display_name}"
+                if workflow.is_async:
+                    display_name += " (Real-time)"
+                workflow_options[display_name] = workflow.name
             
-            if workflows:
-                # Create workflow options for the selectbox
-                workflow_options = {}
-                for workflow in workflows:
-                    display_name = f"{workflow.display_name}"
-                    if workflow.is_async:
-                        display_name += " (Real-time)"
-                    workflow_options[display_name] = workflow.name
+            # Update available workflows in session state
+            st.session_state.available_workflows = list(workflow_options.keys())
+            
+            # Workflow selection
+            selected_workflow_display = st.selectbox(
+                "Select Workflow:",
+                list(workflow_options.keys()),
+                help="Choose a guided workflow for structured execution"
+            )
+            
+            # Update selected workflow in session state
+            if selected_workflow_display:
+                st.session_state.selected_workflow = workflow_options[selected_workflow_display]
                 
-                # Update available workflows in session state
-                st.session_state.available_workflows = list(workflow_options.keys())
-                
-                # Workflow selection
-                selected_workflow_display = st.selectbox(
-                    "Select Workflow:",
-                    list(workflow_options.keys()),
-                    help="Choose a guided workflow for structured execution"
-                )
-                
-                # Update selected workflow in session state
-                if selected_workflow_display:
-                    st.session_state.selected_workflow = workflow_options[selected_workflow_display]
+                # Show workflow description
+                selected_workflow_def = registry.get_workflow(st.session_state.selected_workflow)
+                if selected_workflow_def:
+                    st.caption(f"📋 {selected_workflow_def.description}")
                     
-                    # Show workflow description
-                    selected_workflow_def = registry.get_workflow(st.session_state.selected_workflow)
-                    if selected_workflow_def:
-                        st.caption(f"📋 {selected_workflow_def.description}")
-                        
-                        # Show workflow phases
-                        if selected_workflow_def.phases and selected_workflow_def.phases != ["All"]:
-                            phases_text = ", ".join(selected_workflow_def.phases)
-                            st.caption(f"🔧 Phases: {phases_text}")
-            else:
-                st.warning("No workflows available or workflow system not configured")
-                st.session_state.available_workflows = []
-                st.session_state.selected_workflow = None
-                
-        except Exception as e:
-            logger.error(f"Error loading workflows: {e}", exc_info=True)
-            st.error(f"Error loading workflows: {e}")
+                    # Show workflow phases
+                    if selected_workflow_def.phases and selected_workflow_def.phases != ["All"]:
+                        phases_text = ", ".join(selected_workflow_def.phases)
+                        st.caption(f"🔧 Phases: {phases_text}")
+        else:
+            st.warning("No workflows available or workflow system not configured")
             st.session_state.available_workflows = []
             st.session_state.selected_workflow = None
-    else:
-        # In unguided mode, clear workflow selection
+            
+    except Exception as e:
+        logger.error(f"Error loading workflows: {e}", exc_info=True)
+        st.error(f"Error loading workflows: {e}")
+        st.session_state.available_workflows = []
         st.session_state.selected_workflow = None
 
     # --- PHASE SELECTION ===
@@ -441,7 +429,7 @@ with st.sidebar:
         "Max Actions per Request",
         min_value=1,
         max_value=20,
-        value=MAX_LOOP_ITERATIONS,
+        value=10,
         key="max_loop_iterations",
         help="Maximum number of LLM actions before stopping execution"
     )
@@ -1110,1055 +1098,11 @@ def run_async_workflow_integrated(workflow_name: str, provider: str, model_name:
         # Force a rerun to update the UI back to input mode
         st.rerun()
 
-def run_workflow_execution(workflow_name: str, provider: str, model_name: str, base_sys_prompt: str, user_req_id: str):
-    """Runs a guided workflow execution."""
-    bound_logger = logger.bind(user_request_id=user_req_id)
-    
-    # Track execution timing like unguided mode
-    execution_start_time = time.time()
-    
-    try:
-        # Get workflow registry
-        from nifi_mcp_server.workflows.registry import get_workflow_registry
-        registry = get_workflow_registry()
-        
-        # Create workflow executor
-        executor = registry.create_executor(workflow_name)
-        if not executor:
-            error_msg = f"Failed to create executor for workflow: {workflow_name}"
-            bound_logger.error(error_msg)
-            with st.chat_message("assistant"):
-                st.markdown(f"❌ **Error:** {error_msg}")
-            st.session_state.messages.append({
-                "role": "assistant",
-                "content": f"❌ **Error:** {error_msg}"
-            })
-            return
-        
-        bound_logger.info(f"Starting guided workflow execution: {workflow_name}")
-        
-        # Display workflow start message
-        with st.chat_message("assistant"):
-            st.markdown(f"🚀 **Starting guided workflow:** {workflow_name}")
-            st.markdown("*The workflow will execute multiple LLM iterations with tool calls. Please wait for completion...*")
-        
-        st.session_state.messages.append({
-            "role": "assistant",
-            "content": f"🚀 **Starting guided workflow:** {workflow_name}\n\n*The workflow will execute multiple LLM iterations with tool calls. Please wait for completion...*"
-        })
-        
-        # Prepare execution context - match the unguided mode context construction
-        current_objective = st.session_state.get("current_objective", "")
-        if current_objective and current_objective.strip():
-            effective_system_prompt = f"{base_sys_prompt}\n\n## Current Objective\n{current_objective.strip()}"
-            bound_logger.debug("Objective found, appending to base system prompt for workflow.")
-        else:
-            effective_system_prompt = base_sys_prompt
-            bound_logger.debug("No objective set, using base system prompt for workflow.")
-        
-        # Filter out the workflow startup message to prevent duplication
-        # The startup message we just added should not be passed to the workflow
-        filtered_messages = []
-        for msg in st.session_state.messages:
-            # Skip the workflow startup message we just added
-            if (msg.get("role") == "assistant" and 
-                msg.get("content", "").startswith("🚀 **Starting guided workflow:")):
-                continue
-            filtered_messages.append(msg)
-        
-        context = {
-            "provider": provider,
-            "model_name": model_name,
-            "system_prompt": effective_system_prompt,  # Use the effective system prompt with objective
-            "user_request_id": user_req_id,
-            "messages": filtered_messages,  # Use filtered messages instead of all messages
-            "selected_nifi_server_id": st.session_state.get("selected_nifi_server_id"),
-            "selected_phase": st.session_state.get("selected_phase", "All"),
-            "max_loop_iterations": st.session_state.get("max_loop_iterations", MAX_LOOP_ITERATIONS),
-            "max_tokens_limit": st.session_state.get("max_tokens_limit", 8000),  # Add token limit from UI
-            "auto_prune_history": st.session_state.get("auto_prune_history", False),  # Add pruning setting
-            "current_objective": current_objective  # Also pass the objective separately for workflow use
-        }
-        
-        # Execute workflow with real-time LLM call display
-        with st.spinner(f"🤔 Executing Workflow: {workflow_name}"):
-            result = executor.execute(context)
-        
-        # Calculate execution duration
-        execution_duration = time.time() - execution_start_time
-        
-        # Check for successful execution - executor returns {"status": "success"} not {"success": True}
-        if result and result.get("status") == "success":
-            bound_logger.info("Workflow execution completed successfully")
-            
-            # Extract the actual workflow results from shared state
-            shared_state = result.get("shared_state", {})
-            
-            # --- FIX: Get messages from the correct key ---
-            # The complete message history is now in "final_messages"
-            workflow_messages = shared_state.get("final_messages", [])
-            
-            # DEBUG: Log what we actually received
-            workflow_logger = bound_logger.bind(
-                interface="workflow", 
-                direction="ui_processing",
-                data={
-                    "message_count": len(workflow_messages),
-                    "messages": [
-                        {
-                            "index": i,
-                            "role": msg.get("role", "unknown"),
-                            "has_content": bool(msg.get("content")),
-                            "has_tool_calls": bool(msg.get("tool_calls"))
-                        }
-                        for i, msg in enumerate(workflow_messages)
-                    ]
-                }
-            )
-            workflow_logger.info("Received workflow messages in UI")
-            
-            # Get other details from the "unguided_mimic_result" for the summary
-            workflow_result = shared_state.get("unguided_mimic_result", {})
-            
-            # Calculate token counts correctly by summing from all messages
-            total_tokens_in = 0
-            total_tokens_out = 0
-            
-            # Sum up tokens from all assistant messages in the workflow
-            for msg in workflow_messages:
-                if msg.get("role") == "assistant":
-                    total_tokens_in += msg.get("token_count_in", 0)
-                    total_tokens_out += msg.get("token_count_out", 0)
-            
-            total_tokens = total_tokens_in + total_tokens_out
-            
-            # Get workflow execution details for summary
-            loop_count = workflow_result.get("loop_count", 0)
-            tool_calls_executed = workflow_result.get("tool_calls_executed", 0)
-            max_iterations_reached = workflow_result.get("max_iterations_reached", False)
-            tool_results = workflow_result.get("tool_results", [])
-            
-            # Extract unique tool names that were executed
-            executed_tools = []
-            if tool_results:
-                tool_names = [result.get("tool_name") for result in tool_results if result.get("tool_name")]
-                # Get unique tool names while preserving order
-                seen = set()
-                executed_tools = [name for name in tool_names if name not in seen and not seen.add(name)]
-            
-            # Note: Messages are already displayed in real-time during workflow execution
-            # No need to redisplay or reprocess them here - just keep them as they are
-            
-            # Display workflow completion summary with tokens and duration (like unguided mode)
-            with st.chat_message("assistant"):
-                st.markdown("✅ **Workflow completed successfully**")
-                
-                # Display workflow execution details
-                execution_details = []
-                if loop_count > 0:
-                    execution_details.append(f"{loop_count} iteration{'s' if loop_count != 1 else ''}")
-                if tool_calls_executed > 0:
-                    execution_details.append(f"{tool_calls_executed} tool call{'s' if tool_calls_executed != 1 else ''}")
-                if max_iterations_reached:
-                    execution_details.append("⚠️ max iterations reached")
-                
-                if execution_details:
-                    st.markdown(f"*Executed {', '.join(execution_details)}*")
-                
-                # Display tools that were used
-                if executed_tools:
-                    tools_text = ", ".join(executed_tools)
-                    st.markdown(f"*Tools used: {tools_text}*")
-                
-                # Display workflow summary if available
-                if result.get("message"):
-                    st.markdown(f"*{result['message']}*")
-                
-                # Display token and duration summary like unguided mode
-                if total_tokens > 0:
-                    st.markdown(f"📊 **Workflow Summary:** {total_tokens:,} tokens ({total_tokens_in:,} in, {total_tokens_out:,} out) • {execution_duration:.1f}s")
-            
-            # Check if LLM already provided a comprehensive summary before TASK COMPLETE
-            llm_already_provided_summary = False
-            if st.session_state.messages:
-                # Check only the most recent assistant message (the one we just received)
-                last_assistant_msg = None
-                for msg in reversed(st.session_state.messages):
-                    if msg.get("role") == "assistant" and msg.get("content") and not msg.get("workflow_event"):
-                        last_assistant_msg = msg
-                        break
-                
-                if last_assistant_msg:
-                    content = last_assistant_msg.get("content", "")
-                    
-                    # Check if the message contains "TASK COMPLETE" and has substantial content before it
-                    if "TASK COMPLETE" in content:
-                        # Split on TASK COMPLETE and check if there's substantial content before it
-                        parts = content.split("TASK COMPLETE")
-                        if len(parts) > 1 and len(parts[0].strip()) > 200:  # Substantial content before TASK COMPLETE
-                            llm_already_provided_summary = True
-                            bound_logger.info("LLM already provided comprehensive summary - skipping status report")
-            
-            # Status reporting is now handled by workflows, not the UI
-            
-
-            
-            # Add the workflow completion message to session state
-            completion_message = "✅ **Workflow completed successfully**"
-            
-            # Add execution details to completion message
-            if execution_details:
-                completion_message += f"\n\n*Executed {', '.join(execution_details)}*"
-            
-            # Add tools used to completion message
-            if executed_tools:
-                tools_text = ", ".join(executed_tools)
-                completion_message += f"\n\n*Tools used: {tools_text}*"
-            
-            if result.get("message"):
-                completion_message += f"\n\n*{result['message']}*"
-            if total_tokens > 0:
-                completion_message += f"\n\n📊 **Workflow Summary:** {total_tokens:,} tokens ({total_tokens_in:,} in, {total_tokens_out:,} out) • {execution_duration:.1f}s"
-            
-            st.session_state.messages.append({
-                "role": "assistant",
-                "content": completion_message
-            })
-            
-            # Update session state tracking like unguided mode
-            st.session_state.last_request_duration = execution_duration
-            st.session_state.conversation_total_duration = st.session_state.get("conversation_total_duration", 0) + execution_duration
-            
-            bound_logger.info(f"Workflow completed: {total_tokens} tokens ({total_tokens_in} in, {total_tokens_out} out) in {execution_duration:.1f}s")
-        else:
-            # Handle error case - could be status="error" or no result
-            if result:
-                error_msg = result.get("message", f"Workflow failed with status: {result.get('status', 'unknown')}")
-            else:
-                error_msg = "Workflow execution failed - no result returned"
-            bound_logger.error(f"Workflow execution failed: {error_msg}")
-            with st.chat_message("assistant"):
-                st.markdown(f"❌ **Workflow failed:** {error_msg}")
-            
-            st.session_state.messages.append({
-                "role": "assistant",
-                "content": f"❌ **Workflow failed:** {error_msg}"
-            })
-        
-    except Exception as e:
-        bound_logger.error(f"Error in workflow execution: {e}", exc_info=True)
-        with st.chat_message("assistant"):
-            st.markdown(f"❌ **Workflow execution error:** {e}")
-        
-        st.session_state.messages.append({
-            "role": "assistant",
-            "content": f"❌ **Workflow execution error:** {e}"
-        })
-        
-    finally:
-        # Always reset execution state after workflow completion
-        st.session_state.llm_executing = False
-        st.session_state.stop_requested = False
-        # Force a rerun to update the UI back to input mode
-        st.rerun()
-
-def smart_prune_messages(messages: List[Dict], max_tokens: int, provider: str, model_name: str, tools, logger) -> List[Dict]:
-    """
-    Intelligently prune messages to fit within token limits while maintaining conversation structure.
-    
-    Args:
-        messages: List of conversation messages
-        max_tokens: Maximum allowed tokens
-        provider: LLM provider (e.g., "openai")
-        model_name: Model name for token calculation
-        tools: Available tools for token calculation
-        logger: Logger instance
-        
-    Returns:
-        Pruned message list that maintains valid conversation structure
-    """
-    if len(messages) <= 2:  # System + 1 user message minimum
-        return messages
-    
-    # Calculate current tokens
-    try:
-        token_counter = get_token_counter()
-        current_tokens = token_counter.calculate_input_tokens(messages, provider, model_name, tools)
-        logger.info(f"SMART_PRUNE_DEBUG: Initial tokens: {current_tokens}, Limit: {max_tokens}, Messages: {len(messages)}")
-        if current_tokens <= max_tokens:
-            logger.info("SMART_PRUNE_DEBUG: Already under limit, no pruning needed")
-            return messages  # No pruning needed
-    except Exception as e:
-        logger.error(f"Error calculating tokens during smart pruning: {e}")
-        return messages  # Return original on error
-    
-    # Keep system message (index 0) and work with the rest
-    system_message = messages[0]
-    conversation_messages = messages[1:]
-    
-    # DEBUG: Log message structure
-    logger.info(f"SMART_PRUNE_DEBUG: Message structure analysis:")
-    for i, msg in enumerate(conversation_messages):
-        role = msg.get("role", "unknown")
-        has_tool_calls = "tool_calls" in msg and len(msg.get("tool_calls", [])) > 0
-        tool_call_id = msg.get("tool_call_id", "")
-        logger.info(f"  [{i}] {role}{' (has_tool_calls)' if has_tool_calls else ''}{' (tool_call_id: ' + tool_call_id + ')' if tool_call_id else ''}")
-    
-    # Find safe removal points (complete conversation turns)
-    # We'll build a list of "removal groups" - sets of message indices that can be removed together
-    removal_groups = []
-    i = 0
-    
-    while i < len(conversation_messages):
-        message = conversation_messages[i]
-        role = message.get("role")
-        
-        if role == "user":
-            logger.debug(f"SMART_PRUNE_DEBUG: Processing user message at index {i}")
-            # Start of a potential removal group
-            group_start = i
-            group_end = i  # At least include the user message
-            
-            # Look ahead for the complete turn
-            j = i + 1
-            while j < len(conversation_messages):
-                next_message = conversation_messages[j]
-                next_role = next_message.get("role")
-                
-                if next_role == "assistant":
-                    group_end = j
-                    # Check if this assistant message has tool calls
-                    if "tool_calls" in next_message:
-                        # Must include corresponding tool responses
-                        tool_call_ids = {tc.get("id") for tc in next_message.get("tool_calls", [])}
-                        k = j + 1
-                        while k < len(conversation_messages) and tool_call_ids:
-                            tool_message = conversation_messages[k]
-                            if (tool_message.get("role") == "tool" and 
-                                tool_message.get("tool_call_id") in tool_call_ids):
-                                tool_call_ids.remove(tool_message.get("tool_call_id"))
-                                group_end = k
-                            elif tool_message.get("role") != "tool":
-                                break  # End of tool responses
-                            k += 1
-                    # Don't break here - continue looking for more assistant messages in this turn
-                elif next_role == "tool":
-                    # Orphaned tool message (shouldn't happen, but handle gracefully)
-                    group_end = j
-                elif next_role == "user":
-                    # Found next user message - end of current turn
-                    break
-                else:
-                    break
-                j += 1
-            
-            logger.debug(f"SMART_PRUNE_DEBUG: Turn found at indices {group_start}-{group_end}")
-            
-            # More aggressive pruning: preserve only the most recent complete conversation turn
-            # Count how many complete turns we have
-            total_turns = 0
-            temp_i = 0
-            while temp_i < len(conversation_messages):
-                if conversation_messages[temp_i].get("role") == "user":
-                    total_turns += 1
-                    # Skip to end of this turn
-                    temp_j = temp_i + 1
-                    while temp_j < len(conversation_messages) and conversation_messages[temp_j].get("role") != "user":
-                        temp_j += 1
-                    temp_i = temp_j
-                else:
-                    temp_i += 1
-            
-            # Count which turn this is (1-based)
-            current_turn = 0
-            temp_i = 0
-            while temp_i <= i:
-                if conversation_messages[temp_i].get("role") == "user":
-                    current_turn += 1
-                temp_i += 1
-            
-            logger.debug(f"SMART_PRUNE_DEBUG: Turn {current_turn}/{total_turns}, indices {group_start}-{group_end}")
-            
-            # More aggressive pruning when severely over limit
-            turns_to_preserve = 2  # Default: preserve last 2 turns
-            if current_tokens > max_tokens * 2:  # If more than 2x the limit
-                turns_to_preserve = 1  # Only preserve the last 1 turn
-                logger.debug(f"SMART_PRUNE_DEBUG: Severe token limit exceeded ({current_tokens} > {max_tokens * 2}), preserving only last {turns_to_preserve} turn(s)")
-            else:
-                logger.debug(f"SMART_PRUNE_DEBUG: Standard pruning, preserving last {turns_to_preserve} turn(s)")
-            
-            # Only preserve the specified number of recent complete turns
-            if current_turn <= total_turns - turns_to_preserve:
-                removal_groups.append((group_start, group_end))
-                logger.debug(f"SMART_PRUNE_DEBUG: Added removal group {group_start}-{group_end} (turn {current_turn}/{total_turns})")
-            else:
-                logger.debug(f"SMART_PRUNE_DEBUG: Preserving recent turn {current_turn}/{total_turns} at {group_start}-{group_end}")
-            
-            i = group_end + 1
-        else:
-            # Skip orphaned assistant/tool messages (shouldn't happen in well-formed conversations)
-            logger.debug(f"SMART_PRUNE_DEBUG: Skipping orphaned {role} message at index {i}")
-            i += 1
-    
-    # Remove groups from oldest to newest until we're under the token limit
-    # CRITICAL FIX: Remove one group at a time and recalculate
-    logger.info(f"SMART_PRUNE_DEBUG: Found {len(removal_groups)} removal groups to consider")
-    pruned_messages = [system_message] + conversation_messages.copy()
-    
-    # Process removals one at a time, recalculating groups after each removal
-    while removal_groups:
-        # Take the first (oldest) group
-        group_start, group_end = removal_groups[0]
-        
-        # Calculate current indices (accounting for system message)
-        adjusted_start = group_start + 1  # +1 for system message
-        adjusted_end = group_end + 1      # +1 for system message
-        
-        # Validate indices are still valid
-        if adjusted_start >= len(pruned_messages) or adjusted_end >= len(pruned_messages):
-            logger.warning(f"Invalid indices after previous removals: {adjusted_start}-{adjusted_end}, message count: {len(pruned_messages)}")
-            break
-        
-        # Remove the group
-        removed_messages = pruned_messages[adjusted_start:adjusted_end + 1]
-        pruned_messages = pruned_messages[:adjusted_start] + pruned_messages[adjusted_end + 1:]
-        
-        # Log what was removed
-        removed_roles = [msg.get("role") for msg in removed_messages]
-        logger.info(f"Smart pruning removed message group (roles: {removed_roles}) - indices {adjusted_start}-{adjusted_end}")
-        
-        # CRITICAL FIX: Validate message structure after removal
-        if not validate_message_structure(pruned_messages, logger):
-            logger.warning("Message structure validation failed after removal, reverting...")
-            # Revert the removal
-            pruned_messages = pruned_messages[:adjusted_start] + removed_messages + pruned_messages[adjusted_start:]
-            break
-        
-        # Check if we're now under the limit
-        try:
-            new_tokens = token_counter.calculate_input_tokens(pruned_messages, provider, model_name, tools)
-            logger.debug(f"After removing group: {new_tokens} tokens (limit: {max_tokens})")
-            if new_tokens <= max_tokens:
-                break
-        except Exception as e:
-            logger.error(f"Error calculating tokens after removal: {e}")
-            break
-            
-        # Remove the processed group and recalculate remaining groups
-        removal_groups.pop(0)
-        
-        # Recalculate removal groups for the updated message list
-        # This is necessary because indices have shifted after removal
-        conversation_messages = pruned_messages[1:]  # Exclude system message
-        removal_groups = []
-        i = 0
-        
-        while i < len(conversation_messages):
-            message = conversation_messages[i]
-            role = message.get("role")
-            
-            if role == "user":
-                # Start of a potential removal group
-                group_start = i
-                group_end = i  # At least include the user message
-                
-                # Look ahead for the complete turn
-                j = i + 1
-                while j < len(conversation_messages):
-                    next_message = conversation_messages[j]
-                    next_role = next_message.get("role")
-                    
-                    if next_role == "assistant":
-                        group_end = j
-                        # Check if this assistant message has tool calls
-                        if "tool_calls" in next_message:
-                            # Must include corresponding tool responses
-                            tool_call_ids = {tc.get("id") for tc in next_message.get("tool_calls", [])}
-                            k = j + 1
-                            while k < len(conversation_messages) and tool_call_ids:
-                                tool_message = conversation_messages[k]
-                                if (tool_message.get("role") == "tool" and 
-                                    tool_message.get("tool_call_id") in tool_call_ids):
-                                    tool_call_ids.remove(tool_message.get("tool_call_id"))
-                                    group_end = k
-                                elif tool_message.get("role") != "tool":
-                                    break  # End of tool responses
-                                k += 1
-                            # Don't break here - continue looking for more assistant messages in this turn
-                        # Don't break here - continue looking for more assistant messages in this turn
-                    elif next_role == "tool":
-                        # Orphaned tool message (shouldn't happen, but handle gracefully)
-                        group_end = j
-                    elif next_role == "user":
-                        # Found next user message - end of current turn
-                        break
-                    else:
-                        break
-                    j += 1
-                
-                # More aggressive pruning: preserve only the most recent complete conversation turn
-                # Count how many complete turns we have
-                total_turns = 0
-                temp_i = 0
-                while temp_i < len(conversation_messages):
-                    if conversation_messages[temp_i].get("role") == "user":
-                        total_turns += 1
-                        # Skip to end of this turn
-                        temp_j = temp_i + 1
-                        while temp_j < len(conversation_messages) and conversation_messages[temp_j].get("role") != "user":
-                            temp_j += 1
-                        temp_i = temp_j
-                    else:
-                        temp_i += 1
-                
-                # Count which turn this is (1-based)
-                current_turn = 0
-                temp_i = 0
-                while temp_i <= i:
-                    if conversation_messages[temp_i].get("role") == "user":
-                        current_turn += 1
-                    temp_i += 1
-                
-                # More aggressive pruning when severely over limit (same logic as initial pass)
-                try:
-                    recalc_tokens = token_counter.calculate_input_tokens(pruned_messages, provider, model_name, tools)
-                except Exception:
-                    recalc_tokens = max_tokens * 2  # Assume severe if we can't calculate
-                
-                turns_to_preserve = 2  # Default: preserve last 2 turns
-                if recalc_tokens > max_tokens * 2:  # If more than 2x the limit
-                    turns_to_preserve = 1  # Only preserve the last 1 turn
-                
-                # Only preserve the specified number of recent complete turns
-                if current_turn <= total_turns - turns_to_preserve:
-                    removal_groups.append((group_start, group_end))
-                
-                i = group_end + 1
-            else:
-                # Skip orphaned assistant/tool messages (shouldn't happen in well-formed conversations)
-                logger.debug(f"SMART_PRUNE_DEBUG: Skipping orphaned {role} message at index {i}")
-                i += 1
-    
-    # Final validation and token check
-    if not validate_message_structure(pruned_messages, logger):
-        logger.error("Final message structure validation failed, returning original messages")
-        return messages
-    
-    try:
-        final_tokens = token_counter.calculate_input_tokens(pruned_messages, provider, model_name, tools)
-        logger.info(f"Smart pruning complete: {len(messages)} → {len(pruned_messages)} messages, {current_tokens} → {final_tokens} tokens")
-    except Exception:
-        pass
-    
-    return pruned_messages
 
 
-def validate_message_structure(messages: List[Dict], logger) -> bool:
-    """
-    Validate that the message structure is valid for OpenAI API.
-    
-    Rules:
-    1. Every 'tool' message must be preceded by an 'assistant' message with 'tool_calls'
-    2. No orphaned tool messages
-    3. System message should be first (if present)
-    4. Assistant messages with tool_calls must have all corresponding tool responses
-    
-    Args:
-        messages: List of messages to validate
-        logger: Logger instance
-        
-    Returns:
-        True if structure is valid, False otherwise
-    """
-    if not messages:
-        return True
-    
-    # Track tool calls that need responses
-    pending_tool_calls = set()
-    
-    for i, message in enumerate(messages):
-        role = message.get("role")
-        
-        if role == "system":
-            # System message should be first
-            if i != 0:
-                logger.warning(f"System message found at position {i}, should be at position 0")
-                return False
-                
-        elif role == "assistant":
-            # Check if there are unresolved tool calls from previous assistant message
-            if pending_tool_calls:
-                logger.warning(f"Previous assistant message has unresolved tool calls: {pending_tool_calls}")
-                return False
-            
-            # Clear pending tool calls and check for new ones
-            pending_tool_calls.clear()
-            
-            # Check for new tool calls
-            tool_calls = message.get("tool_calls", [])
-            for tool_call in tool_calls:
-                tool_call_id = tool_call.get("id")
-                if tool_call_id:
-                    pending_tool_calls.add(tool_call_id)
-                    
-        elif role == "tool":
-            # Tool message must have a corresponding tool_call_id
-            tool_call_id = message.get("tool_call_id")
-            if not tool_call_id:
-                logger.warning(f"Tool message at position {i} missing tool_call_id")
-                return False
-                
-            if tool_call_id not in pending_tool_calls:
-                logger.warning(f"Tool message at position {i} has orphaned tool_call_id: {tool_call_id}")
-                return False
-                
-            # Remove this tool call from pending
-            pending_tool_calls.remove(tool_call_id)
-            
-        elif role == "user":
-            # User messages should not appear while tool calls are pending
-            if pending_tool_calls:
-                logger.warning(f"User message while tool calls pending: {pending_tool_calls}")
-                return False
-    
-    # Check if there are any unresolved tool calls at the end
-    if pending_tool_calls:
-        logger.warning(f"Unresolved tool calls at end: {pending_tool_calls}")
-        return False
-    
-    return True
 
-# --- Refactored Execution Loop Function --- 
-def run_execution_loop(provider: str, model_name: str, base_sys_prompt: str, user_req_id: str):
-    """Runs the main LLM interaction loop."""
-    bound_logger = logger.bind(user_request_id=user_req_id)
-    loop_count = 0
-    
-    # Use configurable max iterations from session state
-    max_iterations = st.session_state.get("max_loop_iterations", MAX_LOOP_ITERATIONS)
-    
-    # Initialize request-level tracking
-    request_start_time = time.time()
-    request_tokens_in = 0
-    request_tokens_out = 0
-    
-    try:
-        while loop_count < max_iterations:
-            # Check for stop request at the beginning of each iteration
-            if st.session_state.get("stop_requested", False):
-                bound_logger.info("Stop requested by user. Ending execution loop.")
-                with st.chat_message("assistant"):
-                    st.markdown("🛑 **Execution stopped by user.**")
-                st.session_state.messages.append({
-                    "role": "assistant", 
-                    "content": "🛑 **Execution stopped by user.**"
-                })
-                break
-            
-            loop_count += 1
-            current_loop_logger = bound_logger.bind(loop=loop_count)
-            current_loop_logger.info(f"Starting execution loop iteration {loop_count}")
 
-            # --- Prepare context for LLM call (History and System Prompt) ---
-            # 1. Determine the System Prompt
-            current_objective = st.session_state.get("current_objective", "")
-            if current_objective and current_objective.strip():
-                effective_system_prompt = f"{base_sys_prompt}\n\n## Current Objective\n{current_objective.strip()}"
-                current_loop_logger.debug("Objective found, appending to base system prompt.")
-            else:
-                effective_system_prompt = base_sys_prompt
-                current_loop_logger.debug("No objective set, using base system prompt.")
-                
-            # --- Prepare Tools (BEFORE potential pruning) --- 
-            # Get the selected phase and NiFi server ID
-            current_phase = st.session_state.get("selected_phase", "All")
-            current_nifi_server_id = st.session_state.get("selected_nifi_server_id")
-            if not current_nifi_server_id:
-                st.error("No NiFi server selected. Please select a server from the sidebar.")
-                current_loop_logger.error("Aborting loop: No NiFi server ID found in session state.")
-                break # Stop the loop if no server is selected
-                
-            current_loop_logger = current_loop_logger.bind(nifi_server_id=current_nifi_server_id)
-            current_loop_logger.debug(f"Fetching tools for phase: {current_phase}, Server ID: {current_nifi_server_id}")
-            
-            filtered_raw_tools_list = get_available_tools(
-                phase=current_phase, 
-                user_request_id=user_req_id, 
-                selected_nifi_server_id=current_nifi_server_id
-            )
-            
-            # Use filtered tools directly - ChatManager will handle formatting
-            current_loop_logger.debug(f"Using {len(filtered_raw_tools_list) if filtered_raw_tools_list else 0} pre-filtered tools for provider: {provider}") 
-            # --------------------------------------------------
-                
-            # 2. Initialize LLM context messages (AFTER tool prep, BEFORE pruning)
-            llm_context_messages = [{"role": "system", "content": effective_system_prompt}]
-            
-            # CRITICAL FIX: Validate and clean session messages before adding to context
-            # This prevents sending incomplete tool call sequences to the LLM
-            session_messages = st.session_state.messages.copy()
-            
-            # Validate message structure first
-            if not validate_message_structure(session_messages, current_loop_logger):
-                current_loop_logger.warning("Session messages contain invalid tool call sequences. Cleaning...")
-                
-                # Clean the messages by removing incomplete tool call sequences
-                cleaned_messages = []
-                pending_tool_calls = set()
-                
-                for msg in session_messages:
-                    role = msg.get("role")
-                    
-                    if role == "assistant":
-                        # Check if we have pending tool calls from previous assistant message
-                        if pending_tool_calls:
-                            current_loop_logger.debug(f"Skipping assistant message with unresolved tool calls: {pending_tool_calls}")
-                            continue
-                            
-                        # Check for new tool calls
-                        tool_calls = msg.get("tool_calls", [])
-                        if tool_calls:
-                            # Only include if we can expect tool results to follow
-                            # For now, skip assistant messages with tool calls that don't have results
-                            tool_call_ids = {tc.get("id") for tc in tool_calls if tc.get("id")}
-                            
-                            # Look ahead to see if there are corresponding tool results
-                            has_all_results = True
-                            remaining_messages = session_messages[session_messages.index(msg) + 1:]
-                            found_tool_ids = set()
-                            
-                            for future_msg in remaining_messages:
-                                if future_msg.get("role") == "tool":
-                                    tool_call_id = future_msg.get("tool_call_id")
-                                    if tool_call_id in tool_call_ids:
-                                        found_tool_ids.add(tool_call_id)
-                                elif future_msg.get("role") == "assistant":
-                                    # Stop looking when we hit another assistant message
-                                    break
-                            
-                            if found_tool_ids != tool_call_ids:
-                                current_loop_logger.debug(f"Skipping assistant message with incomplete tool results. Expected: {tool_call_ids}, Found: {found_tool_ids}")
-                                continue
-                            else:
-                                pending_tool_calls = tool_call_ids.copy()
-                        
-                        cleaned_messages.append(msg)
-                        
-                    elif role == "tool":
-                        tool_call_id = msg.get("tool_call_id")
-                        if tool_call_id in pending_tool_calls:
-                            cleaned_messages.append(msg)
-                            pending_tool_calls.remove(tool_call_id)
-                        else:
-                            current_loop_logger.debug(f"Skipping orphaned tool message with id: {tool_call_id}")
-                            
-                    elif role == "user":
-                        # User messages are always safe to include
-                        # But if there are pending tool calls, we need to clear them
-                        if pending_tool_calls:
-                            current_loop_logger.debug(f"Clearing pending tool calls due to user message: {pending_tool_calls}")
-                            pending_tool_calls.clear()
-                        cleaned_messages.append(msg)
-                        
-                    else:
-                        # Other message types (shouldn't happen in our case)
-                        cleaned_messages.append(msg)
-                
-                current_loop_logger.info(f"Message cleaning complete: {len(session_messages)} → {len(cleaned_messages)} messages")
-                llm_context_messages.extend(cleaned_messages)
-            else:
-                # Messages are valid, use them as-is
-                llm_context_messages.extend(session_messages)
-                
-            current_loop_logger.debug(f"Initial llm_context_messages length (incl system prompt): {len(llm_context_messages)}")
 
-            # --- Apply Auto-Pruning (if enabled) --- 
-            if st.session_state.get("auto_prune_history", False):
-                current_loop_logger.info("Auto-pruning enabled. Checking token count...")
-                max_tokens = st.session_state.get("max_tokens_limit", 8000)
-                
-                # ENHANCED DEBUG: Always log current state
-                current_loop_logger.info(f"DEBUG: Message count before pruning check: {len(llm_context_messages)}")
-                current_loop_logger.info(f"DEBUG: Max token limit: {max_tokens}")
-                
-                # CRITICAL FIX: Use smart pruning instead of naive pruning
-                try:
-                    token_counter = get_token_counter()
-                    current_tokens = token_counter.calculate_input_tokens(
-                        llm_context_messages, 
-                        provider, 
-                        model_name, 
-                        tools=filtered_raw_tools_list
-                    )
-                    current_loop_logger.info(f"DEBUG: Current token count (incl. tools): {current_tokens}, Limit: {max_tokens}")
-                    
-                    if current_tokens > max_tokens:
-                        current_loop_logger.info(f"DEBUG: Token limit exceeded ({current_tokens} > {max_tokens}). Applying smart pruning...")
-                        
-                        original_count = len(llm_context_messages)
-                        llm_context_messages = smart_prune_messages(
-                            messages=llm_context_messages,
-                            max_tokens=max_tokens,
-                            provider=provider,
-                            model_name=model_name,
-                            tools=filtered_raw_tools_list,
-                            logger=current_loop_logger
-                        )
-                        
-                        # Enhanced verification
-                        final_count = len(llm_context_messages)
-                        current_loop_logger.info(f"DEBUG: Smart pruning results: {original_count} → {final_count} messages")
-                        
-                        try:
-                            final_tokens = token_counter.calculate_input_tokens(llm_context_messages, provider, model_name, tools=filtered_raw_tools_list)
-                            current_loop_logger.info(f"DEBUG: Smart pruning token results: {current_tokens} → {final_tokens} tokens")
-                            
-                            if final_tokens > max_tokens:
-                                current_loop_logger.warning(f"DEBUG: Smart pruning failed to get under limit! Still at {final_tokens} tokens")
-                            else:
-                                current_loop_logger.info(f"DEBUG: Smart pruning successful! Now at {final_tokens} tokens")
-                                
-                        except Exception as verify_e:
-                            current_loop_logger.warning(f"Could not verify final token count after smart pruning: {verify_e}")
-                    else:
-                        current_loop_logger.info(f"DEBUG: Token count is within limit ({current_tokens} <= {max_tokens}). No pruning needed.")
-                        
-                except Exception as token_calc_err:
-                    current_loop_logger.error(f"Error calculating tokens during pruning: {token_calc_err}", exc_info=True)
-                    st.warning("Error calculating tokens, cannot prune history accurately. Disabling auto-prune for this request.")
-                
-                current_loop_logger.info(f"DEBUG: Final message count after pruning check: {len(llm_context_messages)}")
-            else:
-                current_loop_logger.info("DEBUG: Auto-pruning disabled in UI settings.")
-            # ------------------------------------------
-            
-            # --- LLM Action ID --- # Moved Action ID generation closer to the LLM call
-            llm_action_id = str(uuid.uuid4())
-            current_loop_logger = current_loop_logger.bind(action_id=llm_action_id)
-
-            # --- Call LLM --- 
-            # Note: formatted_tools was prepared BEFORE the pruning loop
-            response_data = None
-            with st.spinner(f"Thinking... (Step {loop_count}/{max_iterations}) / Tokens: ~{current_tokens if 'current_tokens' in locals() else 'N/A'}"):
-                current_loop_logger.info(f"Calling LLM ({provider} - {model_name})...")
-                chat_manager = get_chat_manager()
-                response_data = chat_manager.get_llm_response(
-                    messages=llm_context_messages, 
-                    system_prompt=effective_system_prompt,
-                    provider=provider,
-                    model_name=model_name,
-                    user_request_id=user_req_id,
-                    selected_nifi_server_id=current_nifi_server_id,
-                    tools=filtered_raw_tools_list  # Pass pre-filtered tools directly
-                )
-                if response_data is None:
-                     st.error("LLM response was empty.")
-                     current_loop_logger.error("LLM response data was None.")
-                     break
-                else:
-                    current_loop_logger.info(f"Received response from LLM ({provider}).")
-            # ---------------
-            
-            # --- Process LLM Response --- 
-            llm_content = response_data.get("content")
-            tool_calls = response_data.get("tool_calls")
-            error_message = response_data.get("error")
-            token_count_in = response_data.get("token_count_in", 0)
-            token_count_out = response_data.get("token_count_out", 0)
-            
-            # Accumulate tokens for this request
-            request_tokens_in += token_count_in
-            request_tokens_out += token_count_out
-            
-            if error_message:
-                current_loop_logger.error(f"LLM returned an error: {error_message}")
-                with st.chat_message("assistant"):
-                    st.markdown(f"Sorry, I encountered an error: {error_message}")
-                st.session_state.messages.append({"role": "assistant", "content": f"Sorry, I encountered an error: {error_message}"})
-                break
-                
-            # Add assistant message to history (including tool calls if any)
-            assistant_message_to_add = {
-                "role": "assistant", 
-                "token_count_in": token_count_in, 
-                "token_count_out": token_count_out,
-                "action_id": llm_action_id # Add action_id here
-            }
-            if llm_content: assistant_message_to_add["content"] = llm_content
-            if tool_calls: assistant_message_to_add["tool_calls"] = tool_calls
-            st.session_state.messages.append(assistant_message_to_add)
-            current_loop_logger.debug(f"Appended assistant message: Content={'present' if llm_content else 'absent'}, ToolCalls={len(tool_calls) if tool_calls else 0}")
-
-            # Display assistant message (or tool call placeholder) generated in this loop iteration
-            with st.chat_message("assistant"):
-                if tool_calls:
-                    tool_names = [tc.get('function', {}).get('name', 'unknown') for tc in tool_calls]
-                    if tool_names:
-                         st.markdown(f"⚙️ Calling tool(s): `{', '.join(tool_names)}`...")
-                    else:
-                         st.markdown("_(Tool call with no details)_") # Should be rare
-                elif llm_content:
-                    st.markdown(llm_content)
-                    # Only show copy button if "Prepare Full Conversation For Copy" is active
-                    if st.session_state.get("show_copy_buttons", False):
-                        # Replace st.code with st_copy_to_clipboard here as well
-                        # Use llm_action_id if available, otherwise generate a unique one for the key
-                        copy_key = f"loop_copy_{llm_action_id if 'llm_action_id' in locals() and llm_action_id else uuid.uuid4()}"
-                        st_copy_to_clipboard(llm_content, key=copy_key)
-                else:
-                    st.markdown("_(Assistant provided no content or tool calls)_") # Should be rare
-
-                # Add token counts and action ID caption for this message
-                caption_parts = []
-                if token_count_in or token_count_out:
-                     caption_parts.append(f"Tokens: In={token_count_in}, Out={token_count_out}")
-                if llm_action_id: # Use the ID from this loop
-                     caption_parts.append(f"Action ID: `{llm_action_id}`")
-                if caption_parts:
-                    st.caption(" | ".join(caption_parts))
-
-            # --- Handle Tool Calls (if any) --- 
-            if not tool_calls:
-                current_loop_logger.info("No tool calls requested by LLM. Checking for TASK COMPLETE.")
-                # If no tool calls and LLM thinks task is done, break the loop
-                if llm_content and "TASK COMPLETE" in llm_content:
-                    current_loop_logger.info("TASK COMPLETE detected. Ending execution loop.")
-                    break
-                else:
-                    # If no tool call and task not complete, this iteration might be done (e.g., clarifying question)
-                    # Or it could be an unexpected state. For now, assume the loop should end.
-                    current_loop_logger.info("No tool calls and no TASK COMPLETE. Ending loop for this user request.")
-                    break # Exit loop if no tools called and not explicitly complete
-            else:
-                current_loop_logger.info(f"Processing {len(tool_calls)} tool call(s)...")
-                # Process tool calls one by one
-                for tool_call in tool_calls:
-                    # Check for stop request before each tool call
-                    if st.session_state.get("stop_requested", False):
-                        current_loop_logger.info("Stop requested during tool execution. Breaking tool loop.")
-                        break
-                        
-                     # Decision: Use the llm_action_id for subsequent tool calls in this loop iteration
-                     # This simplifies tracing for a single LLM turn, sacrificing granularity
-                     # between multiple tool calls within the same turn.
-                     # We could generate a unique tool_action_id = str(uuid.uuid4()) here if needed.
-                    # tool_action_id = str(uuid.uuid4())
-                    # tool_loop_logger = current_loop_logger.bind(action_id=tool_action_id) # Bind for tool logging
-                    tool_loop_logger = current_loop_logger # Use the logger already bound with llm_action_id
-                    
-                    tool_id = tool_call.get("id")
-                    function_call = tool_call.get("function")
-                    if not tool_id or not function_call or not isinstance(function_call, dict):
-                        tool_loop_logger.error(f"Skipping invalid tool call structure: {tool_call}")
-                        continue
-                    
-                    function_name = function_call.get("name")
-                    function_args_str = function_call.get("arguments", "{}")
-                    tool_loop_logger.info(f"Executing tool: {function_name} (ID: {tool_id})")
-                    
-                    try:
-                        # Parse arguments
-                        arguments = json.loads(function_args_str)
-                        tool_loop_logger.debug(f"Parsed arguments for {function_name}: {arguments}")
-                        
-                        # Execute the tool using mcp_handler
-                        # Pass user_request_id and the llm_action_id from this loop iteration
-                        # Pass the selected NiFi server ID
-                        tool_result = execute_mcp_tool(
-                            tool_name=function_name, 
-                            params=arguments,
-                            selected_nifi_server_id=current_nifi_server_id, # Pass selected server ID
-                            user_request_id=user_req_id, 
-                            action_id=llm_action_id # Use llm_action_id here
-                        )
-                        
-                        # Handle validation errors gracefully
-                        if isinstance(tool_result, dict) and tool_result.get("validation_error"):
-                            # Display friendly message to user
-                            ui_message = tool_result.get("ui_message", "Adjusting parameters...")
-                            with st.chat_message("assistant"):
-                                st.info(f"⚙️ {ui_message}")
-                            
-                            # But send detailed error to LLM for learning
-                            tool_result_content = json.dumps(tool_result)
-                        else:
-                            # Normal tool result - format for LLM
-                            tool_result_content = json.dumps(tool_result) if tool_result is not None else "null"
-                        
-                        # Format result for the LLM
-                        tool_loop_logger.debug(f"Tool {function_name} execution result: {tool_result_content[:200]}...") # Log snippet
-
-                        # Add tool result message to history for the next LLM iteration
-                        st.session_state.messages.append(
-                            {
-                                "role": "tool",
-                                "tool_call_id": tool_id,
-                                "content": tool_result_content, # Send result back as JSON string
-                                # "name": function_name # OpenAI includes name here, Gemini doesn't use it like this
-                            }
-                        )
-                    except json.JSONDecodeError as json_err:
-                        error_content = f"Error parsing arguments for {function_name}: {json_err}"
-                        tool_loop_logger.error(error_content)
-                        st.error(error_content)
-                        st.session_state.messages.append({"role": "tool", "tool_call_id": tool_id, "content": error_content})
-                    except Exception as tool_err:
-                        error_content = f"Error executing tool {function_name}: {tool_err}"
-                        tool_loop_logger.error(error_content, exc_info=True)
-                        st.error(error_content)
-                        st.session_state.messages.append({"role": "tool", "tool_call_id": tool_id, "content": error_content})
-                # Continue loop to send tool results back to LLM
-                # No break here - loop continues automatically
-        
-        # --- Handle Loop End Conditions ---
-        if loop_count >= max_iterations:
-            bound_logger.warning(f"Execution loop reached maximum iterations ({max_iterations})")
-            
-            # Display user feedback
-            with st.chat_message("assistant"):
-                st.markdown(f"⚠️ **Reached maximum action limit ({max_iterations} actions)**")
-                st.markdown("I've completed as much as possible within the action limit.")
-            
-            # Add visible message to history
-            st.session_state.messages.append({
-                "role": "assistant",
-                "content": f"⚠️ **Reached maximum action limit ({max_iterations} actions)**\n\nI've completed as much as possible within the action limit."
-            })
-            
-            # Status reporting is now handled by workflows, not the UI
-        
-    except Exception as e:
-        bound_logger.error(f"Unexpected error in execution loop: {e}", exc_info=True)
-        st.error(f"An unexpected error occurred: {e}")
-    finally:
-        # Calculate final timing and update session totals
-        request_end_time = time.time()
-        request_duration = request_end_time - request_start_time
-        
-        # Update session state with request totals
-        st.session_state.last_request_tokens_in = request_tokens_in
-        st.session_state.last_request_tokens_out = request_tokens_out
-        st.session_state.last_request_duration = request_duration
-        
-        # Update conversation totals
-        st.session_state.conversation_total_tokens_in += request_tokens_in
-        st.session_state.conversation_total_tokens_out += request_tokens_out
-        st.session_state.conversation_total_duration += request_duration
-        
-        # Display request summary
-        if request_tokens_in > 0 or request_tokens_out > 0:
-            total_request_tokens = request_tokens_in + request_tokens_out
-            with st.chat_message("assistant"):
-                st.markdown(f"📊 **Request Summary:** {total_request_tokens:,} tokens ({request_tokens_in:,} in, {request_tokens_out:,} out) • {request_duration:.1f}s")
-            
-            # Log the summary
-            bound_logger.info(f"Request completed: {total_request_tokens} tokens ({request_tokens_in} in, {request_tokens_out} out) in {request_duration:.1f}s")
-        
-        # Always reset execution state
-        st.session_state.llm_executing = False
-        st.session_state.stop_requested = False
-        # Force a rerun to update the UI back to input mode
-        st.rerun()
 
 # --- User Input Handling --- 
 # Defensive reset: if a previous async workflow completed, ensure input UI returns
@@ -2177,9 +1121,9 @@ if is_executing or pending_execution:
     col1, col2 = st.columns([0.85, 0.15])
     with col1:
         if pending_execution:
-            st.info(f"🚀 Starting LLM execution... (Max {st.session_state.get('max_loop_iterations', MAX_LOOP_ITERATIONS)} actions)")
+            st.info(f"🚀 Starting LLM execution... (Max {st.session_state.get('max_loop_iterations', 10)} actions)")
         else:
-            st.info(f"🤔 LLM is processing... (Max {st.session_state.get('max_loop_iterations', MAX_LOOP_ITERATIONS)} actions)")
+            st.info(f"🤔 LLM is processing... (Max {st.session_state.get('max_loop_iterations', 10)} actions)")
     with col2:
         if st.button("🛑 Stop", key="stop_btn", help="Stop LLM execution", use_container_width=True):
             st.session_state.stop_requested = True
@@ -2199,50 +1143,27 @@ if is_executing or pending_execution:
         # Extract execution parameters
         provider, model_name, base_system_prompt, user_req_id, user_input = pending_execution
         
-        # Choose execution mode based on workflow configuration
-        execution_mode = st.session_state.get("workflow_execution_mode", "unguided")
+        # Simplified execution - only async workflow supported
+        selected_workflow = st.session_state.get("selected_workflow")
+        bound_logger = logger.bind(user_request_id=user_req_id)
+        bound_logger.info(f"Async workflow execution: selected_workflow='{selected_workflow}'")
         
-        if execution_mode == "guided":
-            # Check if a workflow is selected
-            selected_workflow = st.session_state.get("selected_workflow")
-            bound_logger = logger.bind(user_request_id=user_req_id)
-            bound_logger.info(f"Guided mode execution: selected_workflow='{selected_workflow}', execution_mode='{execution_mode}'")
-            
-            if selected_workflow:
-                # Check if this is an async workflow
-                if selected_workflow == "async_unguided_mimic":
-                    # Run async workflow with integrated real-time UI
-                    try:
-                        run_async_workflow_integrated(selected_workflow, provider, model_name, base_system_prompt, user_req_id, user_input)
-                    except Exception as e:
-                        bound_logger.error(f"Failed to run async workflow: {e}", exc_info=True)
-                        st.error(f"Failed to run async workflow: {str(e)}")
-                        # Fall back to regular workflow execution
-                        run_workflow_execution(selected_workflow, provider, model_name, base_system_prompt, user_req_id)
-                else:
-                    # Run regular guided workflow
-                    run_workflow_execution(
-                        workflow_name=selected_workflow,
-                        provider=provider, 
-                        model_name=model_name, 
-                        base_sys_prompt=base_system_prompt, 
-                        user_req_id=user_req_id
-                    )
-            else:
-                # No workflow selected, show error and fall back to unguided
-                bound_logger.warning(f"Guided mode selected but no workflow chosen (selected_workflow='{selected_workflow}'), falling back to unguided")
-                with st.chat_message("assistant"):
-                    st.markdown("⚠️ **No workflow selected.** Falling back to unguided mode.")
-                st.session_state.messages.append({
-                    "role": "assistant",
-                    "content": "⚠️ **No workflow selected.** Falling back to unguided mode."
-                })
-                
-                # Call the traditional execution loop
-                run_execution_loop(provider=provider, model_name=model_name, base_sys_prompt=base_system_prompt, user_req_id=user_req_id)
+        if selected_workflow == "async_unguided_mimic":
+            # Run async workflow with integrated real-time UI
+            try:
+                run_async_workflow_integrated(selected_workflow, provider, model_name, base_system_prompt, user_req_id, user_input)
+            except Exception as e:
+                bound_logger.error(f"Failed to run async workflow: {e}", exc_info=True)
+                st.error(f"Failed to run async workflow: {str(e)}")
         else:
-            # Call the traditional execution loop
-            run_execution_loop(provider=provider, model_name=model_name, base_sys_prompt=base_system_prompt, user_req_id=user_req_id)
+            # No supported workflow selected
+            bound_logger.warning(f"No supported workflow selected: '{selected_workflow}'")
+            with st.chat_message("assistant"):
+                st.markdown("⚠️ **No supported workflow selected.** Only 'async_unguided_mimic' is currently supported.")
+            st.session_state.messages.append({
+                "role": "assistant",
+                "content": "⚠️ **No supported workflow selected.** Only 'async_unguided_mimic' is currently supported."
+            })
 
 else:
     # Show normal input when not executing and no pending execution
